Reorganize batch jobs into specific packages (airflow and spark)
