+++ app-backend/batch/src/main/resources/application.conf
@@ -45,12 +45,12 @@ landsat8 {
 }
 
 export-def {
-  awsRegion = "us-east-1"
-  bucketName = "export-definitions"
+  awsRegion   = "us-east-1"
+  bucketName  = "export-definitions"
   awsDataproc = "dataproc.rasterfoundry.com."
-  sparkJarS3 = "s3://rasterfoundry-global-artifacts-us-east-1/batch"
-  sparkJar = "rf-batch-b933688.jar"
-  sparkClass = "com.azavea.rf.batch.export.Export"
+  sparkJarS3  = "s3://rasterfoundry-global-artifacts-us-east-1/batch"
+  sparkJar    = "rf-batch-b933688.jar"
+  sparkClass  = "com.azavea.rf.batch.export.spark.Export"
   sparkMemory = "2G"
 }
