+++ app-backend/batch/src/main/scala/com/azavea/rf/batch/landsat8/airflow/ImportLandsat8.scala
@@ -1,4 +1,4 @@
-package com.azavea.rf.batch.landsat8
+package com.azavea.rf.batch.landsat8.airflow
 
 import com.azavea.rf.batch.Job
 import com.azavea.rf.batch.util._
