+++ app-backend/batch/src/test/scala/com/azavea/rf/batch/ingest/spark/NoDataPatternSpec.scala
@@ -1,4 +1,4 @@
-package com.azavea.rf.batch.ingest
+package com.azavea.rf.batch.ingest.spark
 
 import com.azavea.rf.batch.ingest.model._
